Show & edit temporary fee fields in feeThresholds section
